modify : modularize the cabinet-flow component
